requirements.txt
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
click==8.1.7
Flask==3.1.2
flatbuffers==25.12.19
geographiclib==2.1
geopy==2.4.1
h3==4.4.1
itsdangerous==2.2.0
Jinja2==3.1.6
jplephem==2.23
MarkupSafe==3.0.3
numpy==2.0.2
pycparser==2.23
pytz==2025.2
setuptools==80.9.0
sgp4==2.25
skyfield==1.53
timezonefinder==8.2.0
Werkzeug==3.1.4
wheel==0.45.1
gunicorn==23.0.0
ics==0.7.2
matplotlib>=3.7.0
google-generativeai==0.8.3
